requirements.txt
langchain
python-dotenv
langchain_community
ipykernel
pypdf
bs4
beautifulsoup4
wikipedia
arxiv
chromadb
pymupdf
langchain-text-splitters
sentence-transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
langchain-groq
streamlit
